refactor: split DiagnosticsList and DiagnosticsTable
