+++ Composer/packages/client/src/pages/diagnostics/DiagnosticsTable.tsx
@@ -1,0 +1,66 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
+import { ScrollablePane, ScrollbarVisibility } from 'office-ui-fabric-react/lib/ScrollablePane';
+import { useMemo, useState } from 'react';
+import formatMessage from 'format-message';
+import { RouteComponentProps } from '@reach/router';
+import { useRecoilValue } from 'recoil';
+import { css } from '@emotion/core';
+
+import { Pagination } from '../../components/Pagination';
+import { diagnosticsSelectorFamily } from '../../recoilModel/selectors/diagnosticsPageSelector';
+
+import { IDiagnosticInfo } from './types';
+import { DiagnosticList } from './DiagnosticList';
+
+// -------------------- Styles -------------------- //
+
+const listRoot = css`
+  position: relative;
+  overflow-y: auto;
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+`;
+
+const tableView = css`
+  position: relative;
+  flex-grow: 1;
+`;
+
+// -------------------- Diagnosticist -------------------- //
+export interface IDiagnosticListProps extends RouteComponentProps {
+  skillId?: string;
+  showType: string;
+  onItemClick: (item: IDiagnosticInfo) => void;
+}
+
+const itemCount = 10;
+
+export const DiagnosticsTable: React.FC<IDiagnosticListProps> = (props) => {
+  const { onItemClick, skillId = '', showType } = props;
+  const diagnostics = useRecoilValue(diagnosticsSelectorFamily(skillId));
+  const availableDiagnostics = showType ? diagnostics.filter((x) => x.severity === showType) : diagnostics;
+  const [pageIndex, setPageIndex] = useState<number>(1);
+
+  const pageCount: number = useMemo(() => {
+    return Math.ceil(availableDiagnostics.length / itemCount) || 1;
+  }, [availableDiagnostics]);
+
+  const showItems = availableDiagnostics.slice((pageIndex - 1) * itemCount, pageIndex * itemCount);
+
+  return (
+    <div css={listRoot} data-testid="diagnostics-table-view" role="main">
+      <div aria-label={formatMessage('Diagnostic list')} css={tableView} role="region">
+        <ScrollablePane scrollbarVisibility={ScrollbarVisibility.auto}>
+          <DiagnosticList diagnosticItems={showItems} skillId={skillId} onItemClick={onItemClick} />
+        </ScrollablePane>
+      </div>
+      <Pagination pageCount={pageCount} onChange={setPageIndex} />
+    </div>
+  );
+};
